Convert from PyInquirer to InquirerPy
PyInquirer requires an old version of prompt_toolkit (1.0.14)
which imports Mapping from collections, which appears to be
deprecated. Mapping should be imported from collections.abc.
This breaks in Python 3.10. InquirerPy uses the newer version
of prompt_toolkit but lacks the Separator class.

This switches the requirement from PyInquirer to InquirerPy and
adds a copy of the Separator class to MainMenu.py.